src/main/java/io/hhplus/tdd/point/UserPointService.java
package io.hhplus.tdd.point;

import io.hhplus.tdd.database.UserPointTable;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class UserPointService {

    private final UserPointTable userPointTable;

    public UserPoint chargePoint(long id, long amount) {
        UserPoint userPoint = userPointTable.selectById(id);
        UserPoint updatedPoint = userPoint.addPoint(amount);
        return userPointTable.insertOrUpdate(userPoint.id(), updatedPoint.point());
    }
}
